+++ tests/test-mathutil-lcm.c
@@ -5,15 +5,15 @@
 #include "utils.h"
 // -----------------------------------------------------------------------------
 int main(void) {
-  int trueResult = 70;
-  int a = 14, b = 35, lcm;
+  long trueResult = 70;
+  long a = 14, b = 35, lcm;
 
   // Get GCD
   calculateLCM(a, b, &lcm);
 
   // Check
   if (lcm != trueResult)
-    printf("Error:\n\tComputed: %d != True: %d\n", lcm, trueResult);
+    printf("Error:\n\tComputed: %ld != True: %ld\n", lcm, trueResult);
 
   return 0;
 }
